4kast.ai-backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic-settings
python-multipart
hdbcli
urllib3
mpmath 
sympy
networkx
MarkupSafe 
filelock
jinja2
pandas>=2.2.3
numpy==1.24.4
statsmodels
scikit-learn
prophet
pmdarima>=2.0.4
xgboost
tensorflow
plotly
matplotlib
streamlit
seaborn